Hackish way to add Projector and Shutter controls
The Draken has GCode for sending the power-on (M605) and power-off (M606)
commands to the projectors it supports (Acer H6510BD and ViewSonic PJD7820HD.

It also has a lens shutter servo with commands to open/close it.

These changes add button controls for these two features, but not in a way
that is either generic or detectible... so not really the best solution yet.